requirements.txt
streamlit>=1.33
pandas>=2.1
numpy>=1.26

# Parsing and NLP
pdfplumber>=0.11
python-docx>=1.1.2
spacy>=3.7
textblob>=0.18
scikit-learn>=1.4
transformers>=4.41
torch>=2.2
sentencepiece>=0.1.99
safetensors>=0.4.3
accelerate>=0.29

# Visualization
plotly>=5.20
matplotlib>=3.8
wordcloud>=1.9.3
networkx>=3.2
kaleido>=0.2.1

# Imaging (for word cloud and PNG handling)
pillow>=10.2

# Google Gemini API for multimodal document processing
google-generativeai>=0.3.0

# Install the spaCy English model at build time (fixes E050 and runtime permission issues on Streamlit Cloud)
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl